requirements.txt
Bottleneck==1.3.5
certifi==2022.9.14
colorama==0.4.5
cycler==0.11.0
Cython==0.29.32
fonttools==4.25.0
gensim==4.1.2
igraph==0.10.0
joblib==1.1.0
kiwisolver==1.4.2
leidenalg==0.8.10
llvmlite==0.38.0
matplotlib==3.5.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
munkres==1.1.4
mysqlclient==2.0.3
neo4j==5.0.1
networkx==2.8.4
numba==0.55.1
numexpr==2.8.3
numpy==1.21.5
packaging==21.3
pandas==1.4.4
Pillow==9.2.0
pip==22.1.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-louvain==0.15
pytz==2022.1
PyYAML==6.0
scikit-learn==1.1.1
scipy==1.7.3
seaborn==0.11.2
setuptools==63.4.1
six==1.16.0
smart-open==5.2.1
texttable==1.6.4
threadpoolctl==2.2.0
tornado==6.2
tqdm==4.64.0
wheel==0.37.1
